+++ tests/Unit/Services/IM/IMServiceBuilderTest.php
@@ -14,7 +14,9 @@ declare(strict_types=1);
 namespace Bitrix24\SDK\Tests\Unit\Services\IM;
 
 use Bitrix24\SDK\Services\IM\Counters\Service\Counters;
+use Bitrix24\SDK\Services\IM\Department\Service\Department;
 use Bitrix24\SDK\Services\IM\Dialog\Service\Dialog;
+use Bitrix24\SDK\Services\IM\Disk\Service\Disk;
 use Bitrix24\SDK\Services\IM\IMServiceBuilder;
 use Bitrix24\SDK\Services\IM\Recent\Service\Recent;
 use Bitrix24\SDK\Services\IM\User\Service\User;
@@ -44,6 +46,12 @@ class IMServiceBuilderTest extends TestCase
         $this::assertSame($this->serviceBuilder->notify(), $this->serviceBuilder->notify());
     }
 
+    public function testGetDiskService(): void
+    {
+        $this->assertInstanceOf(Disk::class, $this->serviceBuilder->disk());
+        $this->assertSame($this->serviceBuilder->disk(), $this->serviceBuilder->disk());
+    }
+
     public function testGetChatService(): void
     {
         $this::assertSame($this->serviceBuilder->chat(), $this->serviceBuilder->chat());
@@ -69,6 +77,12 @@ class IMServiceBuilderTest extends TestCase
     {
         $this->assertInstanceOf(Counters::class, $this->serviceBuilder->counters());
         $this->assertSame($this->serviceBuilder->counters(), $this->serviceBuilder->counters());
+    }
+
+    public function testGetDepartmentService(): void
+    {
+        $this->assertInstanceOf(Department::class, $this->serviceBuilder->department());
+        $this->assertSame($this->serviceBuilder->department(), $this->serviceBuilder->department());
     }
 
     public function testGetUserStatusService(): void